Move DataJsonEncoder and JsonCodecs to the sdk module
