+++ sdk/shared/src/main/scala/org/ergoplatform/sdk/DataJsonEncoder.scala
@@ -1,4 +1,4 @@
-package sigmastate.serialization
+package org.ergoplatform.sdk
 
 import java.math.BigInteger
 import io.circe._
@@ -10,14 +10,18 @@ import scalan.RType
 import scorex.util._
 import sigmastate.Values.{Constant, EvaluatedValue}
 import sigmastate._
-import sigmastate.eval._
 import sigmastate.lang.SigmaParser
+import sigmastate.eval._
 import special.collection.Coll
 import special.sigma._
 import debox.cfor
 import sigmastate.exceptions.SerializerException
 import scala.collection.compat.immutable.ArraySeq
 import scala.collection.mutable
+import fastparse.{Parsed, parse}
+import sigmastate.serialization.SigmaSerializer
+import sigmastate.serialization.DataSerializer
+import sigmastate.serialization.ErgoTreeSerializer
 
 object DataJsonEncoder {
   def encode[T <: SType](v: T#WrappedType, tpe: T): Json = {
